move handleTopContainers to own file
